+++ src/wetterdienst/provider/wsv/pegel/api.py
@@ -271,6 +271,7 @@ class WsvPegelRequest(TimeseriesRequest):
             ttl=CacheExpiry.ONE_HOUR,
             client_kwargs=self.settings.fsspec_client_kwargs,
             cache_disable=self.settings.cache_disable,
+            use_certifi=self.settings.use_certifi,
         )
         file.raise_if_exception()
         df = pl.read_json(