+++ src/wetterdienst/settings.py
@@ -35,6 +35,7 @@ class Settings(BaseSettings):
             "headers": {"User-Agent": f"wetterdienst/{__import__('wetterdienst').__version__} ({platform.system()})"},
         },
     )
+    use_certifi: bool = Field(default=False)
     ts_humanize: bool = True
     ts_shape: Literal["wide", "long"] = "long"
     ts_convert_units: bool = True